+++ CHANGELOG.md
@@ -6,7 +6,8 @@ This package was formerly known as `fmutool`.
 * ADDED: `fmucontainer` support `-sequential` mode
 * ADDED: FMI-3.0 early support. Use `-fmi 3` option for `fmucontainer`
 * CHANGE: `fmu_operations` and `fmu_container` packages are renamed `operations` and `container`
-* CHANGE: `FMUException` class is renamed `FMUError`
+* CHANGE: Introduction of `FMUPort` and `port_attrs` method to replace `scalar_attrs` and `scalar_type`
+* CHANGE: `FMUException` and `OperationException`  classes are renamed `FMUError` and `OperationError`
 
 # Version 1.8.4.2
 * FIXED: `fmucontainer` re-set start values for causality=input after fmi2EnterInitialization as workaround for some FMUs.